# Optimize the polynomial evaluation