--- a/base/macros.hpp
+++ b/base/macros.hpp
@@ -15,22 +15,27 @@ namespace base {
 // TODO(phl): See whether that |COMPILER_MSVC| macro can be removed from port.h.
 #if defined(_MSC_VER) && defined(__clang__)
 #define PRINCIPIA_COMPILER_CLANG_CL 1
+#define PRINCIPIA_VECTORCALL
 char const* const CompilerName = "Clang-cl";
 char const* const CompilerVersion = __VERSION__;
 #elif defined(__clang__)
 #define PRINCIPIA_COMPILER_CLANG 1
+#define PRINCIPIA_VECTORCALL
 char const* const CompilerName = "Clang";
 char const* const CompilerVersion = __VERSION__;
 #elif defined(_MSC_VER)
+#define PRINCIPIA_VECTORCALL __vectorcall
 #define PRINCIPIA_COMPILER_MSVC 1
 char const* const CompilerName = "Microsoft Visual C++";
 char const* const CompilerVersion = STRINGIFY_EXPANSION(_MSC_FULL_VER);
 #elif defined(__ICC) || defined(__INTEL_COMPILER)
 #define PRINCIPIA_COMPILER_ICC 1
+#define PRINCIPIA_VECTORCALL
 char const* const CompilerName = "Intel C++ Compiler";
 char const* const CompilerVersion = __VERSION__;
 #elif defined(__GNUC__)
 #define PRINCIPIA_COMPILER_GCC 1
+#define PRINCIPIA_VECTORCALL
 char const* const CompilerName = "G++";
 char const* const CompilerVersion = __VERSION__;
 #else