--- a/numerics/polynomial_in_monomial_basis_body.hpp
+++ b/numerics/polynomial_in_monomial_basis_body.hpp
@@ -398,7 +398,7 @@ operator()(Argument const argument) const {
 template<typename Value_, typename Argument_, int degree_>
 Derivative<Value_, Argument_>
 PolynomialInMonomialBasis<Value_, Argument_, degree_>::
-EvaluateDerivative(Argument const& argument) const {
+EvaluateDerivative(Argument const argument) const {
   return Evaluator<Value, Difference<Argument>, degree_>::EvaluateDerivative(
       coefficients_, argument - origin_, evaluator_);
 }