--- a/numerics/polynomial_in_чебышёв_basis.hpp
+++ b/numerics/polynomial_in_чебышёв_basis.hpp
@@ -127,8 +127,8 @@ class PolynomialInЧебышёвBasis<Value_, Argument_, degree_>
       default;
 
   Value __vectorcall operator()(Argument argument) const override;
-  Derivative<Value, Argument> EvaluateDerivative(
-      Argument const& argument) const override;
+  Derivative<Value, Argument> __vectorcall EvaluateDerivative(
+      Argument argument) const override;
 
   constexpr int degree() const override;
   bool is_zero() const override;